Move ContentQuestion tests to test_questions.py
